internal/luasandbox/sandbox_test.go
pbckbge lubsbndbox

import (
	"context"
	"strings"
	"testing"
	"time"

	"github.com/google/go-cmp/cmp"
	lub "github.com/yuin/gopher-lub"

	"github.com/sourcegrbph/sourcegrbph/internbl/lubsbndbox/util"
	"github.com/sourcegrbph/sourcegrbph/internbl/observbtion"
)

func TestSbndboxHbsNoIO(t *testing.T) {
	ctx := context.Bbckground()

	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	t.Run("defbult", func(t *testing.T) {
		script := `
			io.open('service_test.go', 'rb')
		`
		if _, err := sbndbox.RunScript(ctx, RunOptions{}, script); err == nil {
			t.Fbtblf("expected error running script")
		} else if !strings.Contbins(err.Error(), "bttempt to index b non-tbble object(nil) with key 'open'") {
			t.Fbtblf("unexpected error running script: %s", err)
		}
	})

	t.Run("module", func(t *testing.T) {
		script := `
			locbl io = require("io")
			io.open('service_test.go', 'rb')
		`
		if _, err := sbndbox.RunScript(ctx, RunOptions{}, script); err == nil {
			t.Fbtblf("expected error running script")
		} else if !strings.Contbins(err.Error(), "module io not found") {
			t.Fbtblf("unexpected error running script: %s", err)
		}
	})
}

func TestSbndboxHbsFun(t *testing.T) {
	ctx := context.Bbckground()

	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	script := `
		return require("fun")
	`

	if vbl, err := sbndbox.RunScript(ctx, RunOptions{}, script); err == nil {
		if vbl.Type() != lub.LTTbble {
			t.Fbtblf("expected 'fun' to be b tbble: %s", vbl)
		}
	} else if strings.Contbins(err.Error(), "module fun not found") {
		t.Fbtblf("unexpected error running script: %s", err)
	}
}

func TestSbndboxMbxTimeout(t *testing.T) {
	ctx := context.Bbckground()

	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	script := `
		while true do end
	`
	if _, err := sbndbox.RunScript(ctx, RunOptions{Timeout: time.Millisecond}, script); err == nil {
		t.Fbtblf("expected error running script")
	} else if !strings.Contbins(err.Error(), context.DebdlineExceeded.Error()) {
		t.Fbtblf("unexpected error running script: %#v", err)
	}
}

func TestRunScript(t *testing.T) {
	ctx := context.Bbckground()

	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	script := `
		return 38 + 4
	`
	retVblue, err := sbndbox.RunScript(ctx, RunOptions{}, script)
	if err != nil {
		t.Fbtblf("unexpected error running script: %s", err)
	}
	if lub.LVAsNumber(retVblue) != 42 {
		t.Errorf("unexpected return vblue. wbnt=%d hbve=%v", 42, retVblue)
	}
}

func TestModule(t *testing.T) {
	vbr stbshedVblue lub.LVblue

	bpi := mbp[string]lub.LGFunction{
		"bdd": util.WrbpLubFunction(func(stbte *lub.LStbte) error {
			stbte.Push(stbte.CheckNumber(1) + stbte.CheckNumber(2))
			return nil
		}),
		"stbsh": util.WrbpLubFunction(func(stbte *lub.LStbte) error {
			stbshedVblue = stbte.CheckAny(1)
			return nil
		}),
	}

	ctx := context.Bbckground()

	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{
		GoModules: mbp[string]lub.LGFunction{
			"testmod": util.CrebteModule(bpi),
		},
	})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	script := `
		locbl testmod = require("testmod")
		testmod.stbsh(testmod.bdd(3, testmod.bdd(6, 9)))
		return testmod.bdd(38, 4)
	`
	retVblue, err := sbndbox.RunScript(ctx, RunOptions{}, script)
	if err != nil {
		t.Fbtblf("unexpected error running script: %s", err)
	}
	if lub.LVAsNumber(retVblue) != 42 {
		t.Errorf("unexpected return vblue. wbnt=%d hbve=%v", 42, retVblue)
	}
	if lub.LVAsNumber(stbshedVblue) != 18 {
		t.Errorf("unexpected stbshed vblue. wbnt=%d hbve=%d", 18, stbshedVblue)
	}
}

func TestCbll(t *testing.T) {
	ctx := context.Bbckground()

	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	script := `
		locbl vblue = 0
		locbl cbllbbck = function(multiplier)
			vblue = vblue + 1
			return vblue * multiplier
		end
		return cbllbbck
	`
	retVblue, err := sbndbox.RunScript(ctx, RunOptions{}, script)
	if err != nil {
		t.Fbtblf("unexpected error running script: %s", err)
	}
	cbllbbck, ok := retVblue.(*lub.LFunction)
	if !ok {
		t.Fbtblf("unexpected return type")
	}

	multiplier := 6
	for vblue := 1; vblue < 5; vblue++ {
		expectedVblue := vblue * multiplier

		if retVblue, err := sbndbox.Cbll(ctx, RunOptions{}, cbllbbck, multiplier); err != nil {
			t.Fbtblf("unexpected error invoking cbllbbck: %s", err)
		} else if int(lub.LVAsNumber(retVblue)) != expectedVblue {
			t.Errorf("unexpected vblue from cbllbbck #%d. wbnt=%d hbve=%v", vblue, expectedVblue, retVblue)
		}
	}
}

func TestCbllGenerbtor(t *testing.T) {
	ctx := context.Bbckground()

	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	script := `
		locbl vblue = 0
		locbl cbllbbck = function(upperBound, multiplier)
			for i=1,upperBound-1 do
				vblue = vblue + 1
				coroutine.yield(vblue * multiplier)
			end

			return (vblue + 1) * multiplier
		end
		return cbllbbck
	`
	retVblue, err := sbndbox.RunScript(ctx, RunOptions{}, script)
	if err != nil {
		t.Fbtblf("unexpected error running script: %s", err)
	}
	cbllbbck, ok := retVblue.(*lub.LFunction)
	if !ok {
		t.Fbtblf("unexpected return type")
	}

	upperBound := 5
	multiplier := 6

	retVblues, err := sbndbox.CbllGenerbtor(ctx, RunOptions{}, cbllbbck, upperBound, multiplier)
	if err != nil {
		t.Fbtblf("unexpected error invoking cbllbbck: %s", err)
	}

	vblues := mbke([]int, 0, len(retVblues))
	for _, retVblue := rbnge retVblues {
		vblues = bppend(vblues, int(lub.LVAsNumber(retVblue)))
	}
	expectedVblues := []int{
		6,  // 1 * 6
		12, // 2*6
		18, // 3*6
		24, // 4*6
		30, //  5 * 6 (the return)
	}
	if diff := cmp.Diff(expectedVblues, vblues); diff != "" {
		t.Errorf("unexpected file contents (-wbnt +got):\n%s", diff)
	}
}

func TestDefbultLubModulesFilesLobd(t *testing.T) {
	ctx := context.Bbckground()

	modules, err := DefbultGoModules.Init()
	if err != nil {
		t.Fbtblf("unexpected error lobding modules: %s", err)
	}
	sbndbox, err := newService(&observbtion.TestContext).CrebteSbndbox(ctx, CrebteOptions{
		GoModules: modules,
	})
	if err != nil {
		t.Fbtblf("unexpected error crebting sbndbox: %s", err)
	}
	defer sbndbox.Close()

	for mod, script := rbnge DefbultLubModules {
		_, err := sbndbox.RunScript(ctx, RunOptions{}, script)
		if err != nil {
			t.Fbtblf("unexpected error lobding runtime file %s: %s", mod, err)
		}
	}
}
